adding @ecoinc/ecoism to project removing contract abis from project
