--- src/contracts/HyperProver.contract.ts
@@ -1,166 +1,0 @@
-export const HyperProverAbi = [
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: '_mailbox',
-        type: 'address',
-      },
-      {
-        internalType: 'address',
-        name: '_inbox',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'constructor',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: '_sender',
-        type: 'address',
-      },
-    ],
-    name: 'UnauthorizedDispatch',
-    type: 'error',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: '_sender',
-        type: 'address',
-      },
-    ],
-    name: 'UnauthorizedHandle',
-    type: 'error',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: 'bytes32',
-        name: '_intentHash',
-        type: 'bytes32',
-      },
-    ],
-    name: 'IntentAlreadyProven',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: 'bytes32',
-        name: '_hash',
-        type: 'bytes32',
-      },
-      {
-        indexed: true,
-        internalType: 'address',
-        name: '_claimant',
-        type: 'address',
-      },
-    ],
-    name: 'IntentProven',
-    type: 'event',
-  },
-  {
-    inputs: [],
-    name: 'INBOX',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'MAILBOX',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'PROOF_TYPE',
-    outputs: [
-      {
-        internalType: 'enum IProver.ProofType',
-        name: '',
-        type: 'uint8',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'getProofType',
-    outputs: [
-      {
-        internalType: 'enum IProver.ProofType',
-        name: '',
-        type: 'uint8',
-      },
-    ],
-    stateMutability: 'pure',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'uint32',
-        name: '',
-        type: 'uint32',
-      },
-      {
-        internalType: 'bytes32',
-        name: '_sender',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'bytes',
-        name: '_messageBody',
-        type: 'bytes',
-      },
-    ],
-    name: 'handle',
-    outputs: [],
-    stateMutability: 'payable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: '',
-        type: 'bytes32',
-      },
-    ],
-    name: 'provenIntents',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-] as const
