+++ src/contracts/index.ts
@@ -1,7 +1,6 @@
 export * from './ERC20.contract'
-export * from './Inbox.contract'
-export * from './IntentSource.contract'
-export * from './Prover.contract'
+export * from './intent-source'
+export * from './prover'
 export * from './SimpleAccount.contract'
 export * from './utils'
 // interfaces
