+++ src/intent/fulfill-intent.service.ts
@@ -7,7 +7,7 @@ import {
   TransactionTargetData,
   UtilsIntentService,
 } from './utils-intent.service'
-import { getERC20Selector, InboxAbi } from '../contracts'
+import { getERC20Selector } from '../contracts'
 import { EcoError } from '../common/errors/eco-error'
 import { EcoLogMessage } from '../common/logging/eco-log-message'
 import { Solver } from '../eco-configs/eco-config.types'
@@ -16,6 +16,7 @@ import { EcoConfigService } from '../eco-configs/eco-config.service'
 import { ProofService } from '../prover/proof.service'
 import { ExecuteSmartWalletArg } from '../transaction/smart-wallets/smart-wallet.types'
 import { KernelAccountClientService } from '../transaction/smart-wallets/kernel/kernel-account-client.service'
+import { InboxAbi } from '@ecoinc/ecoism'
 
 /**
  * This class fulfills an intent by creating the transactions for the intent targets and the fulfill intent transaction.
